refactor: create base collection class
